# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "version": "1.0.0",
        "description": "API for managing users in the system. It allows you to create, retrieve, update, and delete users."
    },
    "paths": {
...
        "/users/refresh-token": {
            "post": {
                "summary": "Refresh user token",
                "operationId": "refreshToken",
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "description": "The refresh token",
                                        "example": "some-refresh-token"
                                    }
                                },
                                "required": [
                                    "refreshToken"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "description": "New JWT token"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid refresh token"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}